Modify 'jpl_calc' to accept a reference plus relative julian day
Since Julian Days have leading "irrelevant" digits, there is a loss
of precision when calculating the offset into an ephemeris data block.

This patch modifies 'jpl_calc' to instead take a reference julian epoch,
and an offset (in days) relative to this.  If the reference precision
matches that of the data block boundaries, then the precision of the
offset controls the resulting position precision.

Note that "jde, rel" would be equivanet to "jde + rel, 0.0" if there
was no loss of precision.